[#11] Make `@middleware` decorator support clients. Convert `.decorators` into a sub-package
